worker/requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==2.0
RPi.GPIO==0.5.11
